+++ tests/testthat/test-timer.R
@@ -71,13 +71,6 @@ test_that("tmr_is_stopped", {
   )
 })
 
-test_that("tmr_reset", {
-  expect_identical(tmr_reset(tmr_timer(1)), tmr_timer())
-  expect_identical(tmr_reset(tmr_timer(1)), tmr_timer())
-  expect_identical(tmr_reset(tmr_timer(1), 2), tmr_timer(2))
-  expect_equal(tmr_reset(tmr_timer(1), 2, TRUE), tmr_timer(2, TRUE))
-})
-
 test_that("tmr_start", {
   tmr <- tmr_timer()
   expect_false(tmr_is_started(tmr))